requirements.txt
-e file:.
pyusb==1.2.1
rich==10.12.0

pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.6.1